script behaviour for macos

+++ setup.sh
@@ -1,7 +1,17 @@
 #!/usr/bin/env bash
 stow . --target=$HOME --ignore=.stowrc --ignore=common --ignore=macos --ignore=bazzite --ignore=DS_Store --ignore=setup.sh
 stow common --target=$HOME/.config --ignore=DS_Store
+
 if [[ "$OSTYPE" == "darwin"* ]]; then
     stow macos --target=$HOME/.config --ignore=DS_Store
-    ln -s ~/.config/ssh/config ~/.ssh/config
+    # check if link or file exists before creating it
+    if [ ! -L ~/.ssh/config ] && [ ! -f ~/.ssh/config ]; then
+        ln -s ~/.config/ssh/config ~/.ssh/config
+    else
+        echo "SSH config link or file already exists, skipping..."
+    fi
+fi
+
+if [[ "$OSTYPE" == "linux-gnu"* ]]; then
+    stow bazzite --target=$HOME/.config --ignore=DS_Store
 fi
